Fix for PostgreSQL migration.
Added a comment to the auto-generated unique index for foreignKeys so we can filter by it in the PgSQL-Schema parse. Needed to not generate useless diffs.
Moved the auto-generation of an index to getAddForeignKeyDDL, so it's also triggered when the table is modified. Added a test to prove the functionality and fixed
PsqlPlatformMigrationTest.